requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
redis==5.0.3
python-socketio==5.11.1
sentry-sdk==1.42.0